kleine fixes

+++ kalender_admin.php
@@ -11,6 +11,10 @@ if ($result == false) {
     exit();
 } else {
     $valid_user_id = $result;
+    if(getUserRolle($valid_user_id) != 2){
+        header("Location: index.php");
+        exit();
+    }
 }
 
 # Nutzernamen des Nutzers feststellen
@@ -19,7 +23,7 @@ $username = current_username($valid_user_id);
 # $_GET leeren
 # empty_get($_SERVER['PHP_SELF']);
 # Termin löschen, wenn der entsprechende Button geklickt wird
-if (isset($_GET['del'])) {
+if (isset($_GET['del']) && $admin) {
     $del_event_id = mysql_real_escape_string($_GET['del']);
     $sql = 'DELETE FROM `events` WHERE id = ?';
     $query = $db->prepare($sql);
